common new_queue: squash -Wsign-conversion warning
`sz` is known ≥ 2 here.

+++ lib/common/utils.c
@@ -33,7 +33,7 @@ nodequeue *new_queue(int sz)
 
     if (sz <= 1)
 	sz = 2;
-    q->head = q->tail = q->store = gv_calloc(sz, sizeof(node_t*));
+    q->head = q->tail = q->store = gv_calloc((size_t)sz, sizeof(node_t*));
     q->limit = q->store + sz;
     return q;
 }
